Serverless function for creating order optimized and deployed (after several hours of suffering)
