+++ webapp/firebase.json
@@ -18,6 +18,7 @@
     ]
   },
   "functions": {
-    "source": "functions"
+    "source": "functions",
+    "predeploy": "npm --prefix \"$RESOURCE_DIR\" run build"
   }
 }
